Move Serde trait implementation macro to `serialization` module
